Refactor timetable manager to store
