+++ web/client/src/main.ts
@@ -2,11 +2,13 @@ import { createApp } from 'vue'
 import './style.ts'
 import App from './App.vue'
 import { InfrastructureStore, InfrastructureNameSpace } from './stores/infrastructure-store.js'
+import { TimetableStore, TimetableNamespace } from './stores/timetable-store.js'
 import { createStore } from "vuex";
 
 const store = createStore({
     modules: {
-        [InfrastructureNameSpace]: InfrastructureStore
+        [InfrastructureNameSpace]: InfrastructureStore,
+        [TimetableNamespace]: TimetableStore,
     }
 })
 const app = createApp(App);
